[CSL-2202] Rework Wallet storage effects
This commit renames MonadWalletWebDb into WalletDbReader
and introduce the WalletDbWriter typeclass to model functions
which can only write the database on disk.

This is the first stepping stone into a single "ala IORef" API,
where we are allowed to read multiple times but we can write with
a single API which takes the old state as input, yielding the new
one.